Alert, not working since last commit

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, Text, FlatList } from 'react-native';
+import { View, Text, FlatList, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -34,10 +34,26 @@ class Favorites extends Component {
             {
                text: 'Delete',
                type: 'delete',
-               onPress: () => this.props.deleteFavorite(item.id)
+               onPress: () => {
+                  Alert.alert(
+                     'Delete Favorite?',
+                     'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                     [
+                        { 
+                           text: 'Cancel', 
+                           onPress: () => console.log(item.name + 'Not Deleted'),
+                           style: 'cancel'
+                        },
+                        {
+                           text: 'OK',
+                           onPress: () => this.props.deleteFavorite(item.id)
+                        }
+                     ],
+                     { cancelable: true }
+                  )
+               }
             }
          ]
-
          return(
             <Swipeout right={rightButton} autoClose={true}>
                <ListItem 
@@ -61,7 +77,7 @@ class Favorites extends Component {
          return(
             <View>
                <Text>{this.props.dishes.errMess}</Text>
-                  </View>
+            </View>
          )
       }
       else {
@@ -76,7 +92,6 @@ class Favorites extends Component {
          )
       }
    }
-
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(Favorites);
